+++ src/Responses/AuthSuccess.php
@@ -4,7 +4,7 @@ namespace Saritasa\Laravel\Controllers\Responses;
 
 use Saritasa\Transformers\DtoModel;
 
-class AuthSuccessDTO extends DtoModel {
+class AuthSuccess extends DtoModel {
 
     /**
      * Authentication token